requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
idna==3.6
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.2.0
pycparser==2.21
pypdfium2==4.28.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
tqdm==4.66.2
tzdata==2024.1
urllib3==2.2.1
